chore: Renaming file name to have all math_operations in sorted directly
