+++ tests/unit_tests/base_modules/linear_test.py
@@ -4,7 +4,7 @@ import torch
 from torch.nn.parameter import Parameter
 
 from elasticai.creator.base_modules.linear import Linear, MathOperations
-from elasticai.creator.base_modules.torch_math_operations import TorchMathOperations
+from elasticai.creator.base_modules.math_operations_torch import TorchMathOperations
 from tests.tensor_test_case import TensorTestCase
