+++ tests/unit_tests/base_modules/lstm_test.py
@@ -5,7 +5,7 @@ import torch
 
 from elasticai.creator.base_modules.lstm import LSTM
 from elasticai.creator.base_modules.lstm_cell import LSTMCell
-from elasticai.creator.base_modules.torch_math_operations import TorchMathOperations
+from elasticai.creator.base_modules.math_operations_torch import TorchMathOperations
 from tests.tensor_test_case import TensorTestCase
